Restructure project files and dir
- move json to root folder
- make helper organize
- wip analytic
