+++ package.json
@@ -8,8 +8,7 @@
     "start": "web-ext run --start-url https://tradingview.com/symbols/MYX-FRONTKN/",
     "start:chrome": "web-ext run --target chromium --start-url https://tradingview.com/symbols/MYX-FRONTKN/",
     "build": "web-ext build --overwrite-dest --ignore-files update-data docs *.md package.json package-lock.json",
-    "update-data": "node ./update-data",
-    "scrapping": "node ./update-data/scrapping.js"
+    "update-data": "node update-data"
   },
   "engines": {
     "node": ">=12.14.0",
