+++ src/browser-polyfill.js
