+++ src/page/chart.js
@@ -1,9 +1,8 @@
-/* global TRADING_VIEW_MYR */
-/* global lookForShariah addStaticSyariahIcon isSyariahIconExist deleteSyariahIcon createIcon */
+/* global tsi */
 
 browser.runtime.onMessage.addListener(receiveSignalFromBgScript)
 
-addStaticSyariahIcon()
+tsi.addStaticSyariahIcon()
 
 function getSymbolsFromTitle() {
   const domTittleName = document.getElementsByTagName('title')[0].innerText
@@ -11,11 +10,11 @@ function getSymbolsFromTitle() {
 }
 
 function receiveSignalFromBgScript() {
-  const { s: isShariah } = lookForShariah(`${ TRADING_VIEW_MYR }:${ getSymbolsFromTitle() }`)
+  const { s: isShariah } = tsi.lookForShariah(`${ tsi.TRADING_VIEW_MYR }:${ getSymbolsFromTitle() }`)
 
   if (!isShariah) {
     // didnt found symbol within malaysian stocks
-    deleteSyariahIcon()
+    tsi.deleteSyariahIcon()
     return
   }
 
@@ -23,13 +22,13 @@ function receiveSignalFromBgScript() {
     const element = document.querySelector('[data-name="legend-source-title"]')
 
     // if icon already exist dont do anything
-    if (isSyariahIconExist(element.parentElement)) {
+    if (tsi.isSyariahIconExist(element.parentElement)) {
       return
     }
 
-    element.parentElement.prepend(createIcon({ width: 15, height: 15 }))
+    element.parentElement.prepend(tsi.createIcon({ width: 15, height: 15 }))
   } else {
     // if not syariah delete all icon
-    deleteSyariahIcon()
+    tsi.deleteSyariahIcon()
   }
 }
